Removing email popup
