--- cfgov/jinja2/v1/_includes/organisms/email-popup.html
@@ -1,60 +1,0 @@
-{# ==========================================================================
-
-   Email Pop-up Organism
-
-   ==========================================================================
-
-   Description:
-
-   Sidebar popup that fades in upon scrolling on the Owning a Home sublanding
-   page at /owning-a-home/.
-
-   Creates Pop-up organism:
-
-   ========================================================================== #}
-
-<div class="o-email-popup u-alpha-transition u-alpha-0">
-  <div class="o-email-popup_header u-clearfix">
-    <div class="close">
-      <a>Close <span class="cf-icon cf-icon-delete-round"></span></a>
-    </div>
-  </div>
-  <div class="o-email-popup_body">
-    <div class="o-email-signup">
-      <h2>Buying a home?</h2>
-      <form id="{{ ('o-email-signup_' ~ range(1, 100) | random) }}"
-              class="o-form o-form__email-signup u-clearfix"
-                action="/subscriptions/new/"
-                method="POST"
-                enctype="application/x-www-form-urlencoded"
-                novalidate>
-              <p>Sign up for email tips and info to help you through the process.</p>
-              <div class="m-form-field-with-button" data-qa-hook="formfield-with-button">
-                <div class="form-group">
-                  <input id="form_18"
-                         type="email"
-                         placeholder="Enter your email address"
-                         name="email"
-                         class="m-form-field-with-button_field"
-                         required>
-                </div>
-                <p>
-                  <a href="/owning-a-home/privacy-act-statement/" target="_blank">
-                    View Privacy Act statement
-                  </a>
-                  <br/>
-                </p>
-                <input class="a-btn a-btn__full-on-xs" type="submit" value="Sign up">
-              </div>
-              <div class="form-group">
-                <input type="hidden" name="code" value="USCFPB_127">
-              </div>
-          </form>
-          <div class="o-email-signup_footer">
-            {% import 'molecules/notification.html' as notification with context %}
-            {{ notification.render('success', false, '') }}
-          </div>
-      </div>
-    </div>
-  <div class="o-email-popup_footer"></div>
-</div>
