+++ cfgov/jinja2/v1/_includes/organisms/search-bar.html
@@ -49,7 +49,7 @@
         <div class="o-search-bar_input o-search-bar_col">
             <div class="input-with-btn input-with-btn__bar">
                 <div class="input-with-btn_input {% if value.autocomplete %} m-autocomplete{% endif %}">
-                    <input type="text" name="q" id="q" >
+                    <input class="a-text-input" type="text" name="q" id="q" >
                 </div>
                 <div class="input-with-btn_btn">
                     <button class="a-btn" name="btnSearch" id="btnSearch" type="submit">
