+++ cfgov/jinja2/v1/sublanding-page/index.html
@@ -53,10 +53,6 @@
             {{ render_block.render(block, loop.index) }}
         {% endif %}
     {%- endfor %}
-
-    {% if flag_enabled('EMAIL-POPUP', request) %}
-        {% include 'organisms/email-popup.html' ignore missing %}
-    {% endif %}
 {% endblock %}
 
 {% block content_sidebar_modifiers -%}
