+++ cfgov/unprocessed/css/main.less
@@ -145,7 +145,6 @@
 @import (less) "organisms/sidebar-content.less";
 @import (less) "organisms/snippet-list.less";
 @import (less) "organisms/well.less";
-@import (less) "organisms/email-popup.less";
 
 /* Template pieces
    ========================================================================== */
