+++ cfgov/unprocessed/css/molecules/autocomplete.less
@@ -1,6 +1,10 @@
 
 .m-autocomplete {
     position: relative;
+
+    input {
+        width: 100%;
+    }
 }
 
 .m-autocomplete_results {
