--- cfgov/unprocessed/css/organisms/email-popup.less
@@ -1,156 +1,0 @@
-
-.o-email-popup {
-  width: 270px;
-  position: fixed;
-  bottom: 1em;
-  right: 1em;
-  z-index: 2000;
-  background-color: white;
-  border: 1px solid @gray-20;
-
-  .respond-to-max( @bp-xs-max, {
-    margin-left: -135px;
-    left: 50%;
-  });
-
-  .respond-to-min( @bp-sm-min, {
-    width: 555px;
-  });
-
-  &__visible {
-    display: block;
-  }
-
-  &_header {
-    margin: 0 unit( (@grid_gutter-width / 2) / @base-font-size-px, em);
-    height: 40px;
-
-    .respond-to-min( @bp-sm-min, {
-      margin: 0
-              unit( (@grid_gutter-width / 2) / @base-font-size-px, em)
-              0
-              unit(@grid_gutter-width / @base-font-size-px, em);
-      height: 26px;
-    });
-
-    .close {
-      position: absolute;
-      top: unit( (@grid_gutter-width / 3) / @base-font-size-px, em);
-      right: unit( (@grid_gutter-width / 2) / @base-font-size-px, em);
-      cursor: pointer;
-    }
-  }
-
-  &_body {
-    margin: 0
-            unit( (@grid_gutter-width / 2) / @base-font-size-px, em)
-            unit( (@grid_gutter-width / 2) / @base-font-size-px, em);
-
-    h2 {
-      .h4();
-      margin-bottom:  unit( (@grid_gutter-width / 2) / @font-size, em);
-    }
-
-    .o-email-signup {
-      width: 100% !important;
-    }
-
-    .o-email-signup.form-submitted {
-      margin-bottom: unit( (@grid_gutter-width * 2) / @base-font-size-px, em);
-    }
-
-    input[type="email"] {
-      width: 100%;
-      margin-bottom: unit( (@grid_gutter-width / 2) / @base-font-size-px, em);
-    }
-
-    .respond-to-max( @bp-xs-max, {
-      background: none;
-
-      .o-email-signup.form-submitted {
-        margin-bottom: unit( (@grid_gutter-width * 3) / @base-font-size-px, em);
-      }
-    });
-
-    .respond-to-min( @bp-sm-min, {
-
-      margin: 0
-              unit( @grid_gutter-width  / @base-font-size-px, em)
-              unit( @grid_gutter-width  / @base-font-size-px, em);
-
-      padding-right: 155px;
-      background: url( '../img/modal-house.png' )
-                  no-repeat
-                  right
-                  20px;
-      background-size: 130px;
-
-      h2 {
-        .h3();
-        margin-bottom: unit( (@grid_gutter-width / 3) / @font-size, em);
-      }
-
-      .m-form-field-with-button {
-        p {
-          float: right;
-          position: relative;
-          top: 0.7em;
-        }
-        .a-btn {
-          float: left;
-        }
-      }
-    });
-  }
-
-  .o-email-signup_footer {
-    display: table;
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    margin-top: unit(@grid_gutter-width / @base-font-size-px, em);
-
-    .respond-to-min( @bp-sm-min, {
-      margin-top: unit( @grid_gutter-width / @base-font-size-px, em);
-    });
-
-    .m-notification {
-      width: 100%;
-      text-align: left;
-      vertical-align: middle;
-      border-width: 0;
-
-      &__visible {
-        display: table-cell;
-        padding: unit( (@grid_gutter-width / 3) / @base-font-size-px, em)
-                 unit( (@grid_gutter-width / 2) / @base-font-size-px, em);
-
-        .respond-to-min( @bp-sm-min, {
-          padding: unit( (@grid_gutter-width / 3) / @base-font-size-px, em)
-                   unit( @grid_gutter-width / @base-font-size-px, em);
-        });
-      }
-
-      &_icon {
-        top: 11px;
-        .respond-to-min( @bp-sm-min, {
-            left: unit( @grid_gutter-width / @base-font-size-px, em);
-        });
-      }
-
-      &_content {
-        font-family: "AvenirNextLTW01-Medium",Arial,sans-serif;
-        font-weight:500;
-        text-align: left;
-        padding-left: unit( @grid_gutter-width / @base-font-size-px, em);
-
-        p.h4 {
-          font-size: unit( @base-font-size-px / 16 * 100, % ) !important;
-        }
-      }
-
-
-    }
-  }
-}
