+++ cfgov/unprocessed/js/routes/on-demand/email-signup.js
@@ -6,7 +6,6 @@
 
 var FormSubmit = require( '../../organisms/FormSubmit.js' );
 var validators = require( '../../modules/util/validators' );
-var emailHelpers = require( '../../modules/util/email-popup-helpers' );
 
 var BASE_CLASS = 'o-email-signup';
 
@@ -29,7 +28,4 @@ formSubmit.addEventListener( 'success', function onEmailSignupSuccess( event ) {
   var form = event.form;
   var input = form.querySelector( 'input[name="code"]' );
   var code = input.value;
-  if ( code === 'USCFPB_127' ) {
-    emailHelpers.recordEmailRegistration();
-  }
 } );
